// FrontEnd/Point-Counter/src/app/models/player.model.ts
export interface Player {
  id: number;
  name: string;
  age: number;
  address: string;
  phone: string;
}

export interface CreatePlayer {
  name: string;
  age: number;
  address: string;
  phone: string;
}

export interface UpdatePlayer{
  name: string;
  age: number;
  address: string;
  phone: string;
}